+++ src/app/portal/admin-perfil/admin-conocimiento/admin-conocimiento.component.ts
@@ -16,12 +16,19 @@ export class AdminConocimientoComponent implements OnInit {
   public conocimiento: Conocimiento;
   public categoriasConocimiento: any[];
   public tiposConocimiento: TipoConocimiento[];
+  public tipoConocimientoSeleccionado: TipoConocimiento;
 
   constructor(public activeModal: NgbActiveModal, public conocimientoService: ConocimientoService, public toastService: ToastService) { }
 
   ngOnInit(): void {
     this.cargarCategoriasConocimiento();
     this.cargarTiposConocimiento();
+
+    this.tipoConocimientoSeleccionado = new TipoConocimiento({ id: 0, nombre: "" });
+    if (this.conocimiento.tipo && this.conocimiento.tipo.id) {
+      this.tipoConocimientoSeleccionado = new TipoConocimiento(this.conocimiento.tipo)
+    }
+    console.log(this.tipoConocimientoSeleccionado);
   }
 
   cargarCategoriasConocimiento() {
@@ -31,7 +38,7 @@ export class AdminConocimientoComponent implements OnInit {
         this.categoriasConocimiento.forEach(item => {
           item.isChecked = false;
           this.conocimiento.categorias.forEach(item2 => {
-            if(item.id === item2.id){
+            if (item.id === item2.id) {
               item.isChecked = true;
             }
           });
@@ -54,25 +61,31 @@ export class AdminConocimientoComponent implements OnInit {
     );
   }
 
+  compareTipoConocimiento(tc1: any, tc2: any): boolean {
+    return tc1.id === tc2.id;
+  }
+
   onChangeCategoria(categoria: any, isChecked: boolean) {
-    let categoriaConocimiento: CategoriaConocimiento = new CategoriaConocimiento(categoria);
     if (isChecked) {
-      this.conocimiento.categorias.push(categoriaConocimiento);
+      let tmpCategoria = new CategoriaConocimiento(categoria);
+      this.conocimiento.categorias.push(tmpCategoria);
     } else {
-      let index = this.conocimiento.categorias.findIndex((item)=> item.id === categoriaConocimiento.id);
-      this.conocimiento.categorias.splice(index,1);
+      let index = this.conocimiento.categorias.findIndex((item) => item.id === categoria.id);
+      this.conocimiento.categorias.splice(index, 1);
     }
+    console.log(this.tipoConocimientoSeleccionado);
   }
 
   guardarConocimiento() {
+    this.conocimiento.tipo = new TipoConocimiento(this.tipoConocimientoSeleccionado);
     console.log(this.conocimiento);
-    if(this.conocimiento.id && this.conocimiento.id > 0){
+    if (this.conocimiento.id && this.conocimiento.id > 0) {
       this.conocimientoService.updateConocimiento(this.conocimiento).subscribe(
         data => {
           this.toastService.showSuccess("Conocimiento Actualizado");
           console.log(data);
-          this.activeModal.dismiss("Update OK");
-        }, 
+          this.activeModal.dismiss(this.conocimiento);
+        },
         err => {
           err.error.details.forEach(detail => {
             this.toastService.showDanger(detail);
@@ -84,8 +97,8 @@ export class AdminConocimientoComponent implements OnInit {
         data => {
           this.toastService.showSuccess("Conocimiento Creado");
           console.log(data);
-          this.activeModal.dismiss("Save OK");
-        }, 
+          this.activeModal.dismiss(this.conocimiento);
+        },
         err => {
           err.error.details.forEach(detail => {
             this.toastService.showDanger(detail);